yaRab

+++ src/app/Component/core/AdminSide/products/product-tabel/product-tabel.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Subscription } from 'rxjs';
 import { IproductShow } from 'src/app/models/i-product-variant';
 import { ProductFormService } from 'src/app/services/product-form.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-product-tabel',
@@ -33,12 +34,22 @@ export class ProductTabelComponent implements OnInit {
   }
 
   deleteProduct (id : number) {
-    this.prodServ.deleteProduct(id).subscribe({
-      next: ()=> {this.getallProduct();},
-      error: (e) => {
-        console.log("ERROR When delete product" + e);
-      }
-    })
+    Swal.fire({
+      title: 'هل أنت متأكد من أنك تريد حذف المنتج؟!',
+      confirmButtonColor: '#198754', // لون زر التأكيد
+      confirmButtonText: 'تأكيد',
+      cancelButtonText: 'إلغاء',
+      showCancelButton: true,
+      cancelButtonColor: '#dc3545' // لون زر الإلغاء
+    }).then((result) => {if (result.isConfirmed) {
+      this.prodServ.deleteProduct(id).subscribe({
+        next: ()=> {this.getallProduct();},
+        error: (e) => {
+          console.log("ERROR When delete product" + e);
+        }
+      })
+    }})
+
   }
